+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ features.
 - Added :func:`brownie.functional.bind_arguments`.
 - Allow using :func:`brownie.itools.unique` with non-hashable items.
 - Added :mod:`pickle` support to :class:`brownie.datastructures.LazyList`.
+- Made :class:`brownie.datastructures.ImmutableDict` hashable.
 
 0.4.1
 -----